+++ Project1/ET.h
@@ -9,6 +9,6 @@ class ET : public Unit
 public:
 	ET();
 
-	ET(int id, double tj, int h, int p, int ac);
+	ET(int id, int tj, int h, int p, int ac);
 	void Attack();
 };
